+++ udata/core/reports/api.py
@@ -1,11 +1,12 @@
+import mongoengine
 from flask import request
 from flask_login import current_user
-import mongoengine
 
-from udata.api import api, API, fields
+from udata.api import API, api, fields
 from udata.api_fields import patch
-from .models import Report
+
 from .constants import reports_reasons_translations
+from .models import Report
 
 ns = api.namespace('reports', 'User reported objects related operations (beta)')
 
@@ -38,6 +39,6 @@ class ReportsAPI(API):
 @ns.route('/reasons/', endpoint='reports_reasons')
 class ReportsReasonsAPI(API):
     @api.doc('list_reports_reasons')
-    @ns.response(200, "dictionnary of available reasons associated with their labels", fields.Raw)
+    @ns.response(200, "list of available reasons associated with their labels", fields.Raw)
     def get(self):
         return reports_reasons_translations()