+++ udata/tests/api/test_dataservices_api.py
@@ -1,4 +1,3 @@
-from pprint import pprint
 from flask import url_for
 
 from udata.core.dataservices.models import Dataservice